% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{final_character}
\alias{final_character}
\title{final_character, the example's final signature}
\format{
An object of class \code{character} of length 2.
}
\usage{
final_character
}
\description{
final_character, a potential marker for cancer prognostic and immunotherapy, generated by 'get_final_signature'
}
\keyword{datasets}
