% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene_path}
\alias{gene_path}
\title{gene_path, the pathways geneset}
\format{
An object of class \code{list} of length 27.
}
\usage{
gene_path
}
\description{
gene_path, a list of KEGG Non-metabolic pathways geneset.
}
\keyword{datasets}
