% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genesmbol}
\alias{genesmbol}
\title{genesmbol, the coding genes' length}
\format{
An object of class \code{list} of length 34931.
}
\usage{
genesmbol
}
\description{
genesmbol,a list of coding genes' length, generated by 'get_gene_length'.
}
\keyword{datasets}
