% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_gene_length}
\alias{get_gene_length}
\title{Extract coding genes' length from gene transfer format(GTF) file.}
\usage{
get_gene_length(filepath)
}
\arguments{
\item{filepath}{Input gene transfer format file. It must be an absolute path or the name relatived to the current working directory.}
}
\value{
Return a list of genes' length.
}
\description{
Extract coding genes' length from gene transfer format(GTF) file.
}
