% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mut_matrix}
\alias{mut_matrix}
\title{mut_matrix, mutations matrix}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 673 rows and 35 columns.
}
\usage{
mut_matrix
}
\description{
mut_matrix, the mutations matrix,generated by `get_mut_matrix`
}
\keyword{datasets}
