% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PTMB_matrix}
\alias{PTMB_matrix}
\title{PTMB_matrix, the Pathway-based Tumor Mutational Burden matrix}
\format{
An object of class \code{data.frame} with 27 rows and 35 columns.
}
\usage{
PTMB_matrix
}
\description{
PTMB_matrix, the pathway tumor mutation burden matrix,generated by`get_PTMB`
}
\keyword{datasets}
