% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sur}
\alias{sur}
\title{sur, the samples' survival data}
\format{
An object of class \code{data.frame} with 110 rows and 2 columns.
}
\usage{
sur
}
\description{
sur, a nx2 data frame, the samples' survival data
}
\keyword{datasets}
